refactor: Nested drop down in input-choice
