+++ presentation/component/damage-definition-list/damage-definition-list-item-viewmodel.mjs
@@ -4,7 +4,7 @@ import { isDefined } from "../../../business/util/validation-utility.mjs";
 import { TEMPLATES } from "../../templatePreloader.mjs";
 import ButtonViewModel from "../button/button-viewmodel.mjs";
 import ChoiceAdapter from "../input-choice/choice-adapter.mjs";
-import InputDropDownViewModel from "../input-dropdown/input-dropdown-viewmodel.mjs";
+import InputDropDownViewModel from "../input-choice/input-dropdown/input-dropdown-viewmodel.mjs";
 import InputTextFieldViewModel from "../input-textfield/input-textfield-viewmodel.mjs";
 import InputViewModel from "../../view-model/input-view-model.mjs";
 import DamageAndType from "../../../business/ruleset/skill/damage-and-type.mjs";
