+++ presentation/component/input-choice/input-dropdown/input-dropdown-viewmodel.mjs
@@ -1,6 +1,6 @@
-import { validateOrThrow } from "../../../business/util/validation-utility.mjs";
-import { TEMPLATES } from "../../templatePreloader.mjs";
-import InputChoiceViewModel from "../input-choice/input-choice-viewmodel.mjs";
+import { validateOrThrow } from "../../../../business/util/validation-utility.mjs";
+import { TEMPLATES } from "../../../templatePreloader.mjs";
+import InputChoiceViewModel from "../input-choice-viewmodel.mjs";
 
 /**
  * Represents a drop-down input. The user can select one of a defined list of options.
