+++ presentation/templatePreloader.mjs
@@ -35,7 +35,7 @@ export const TEMPLATES = {
   COMPONENT_LABELED_CONTROL: `${basePathComponent}/labeled-control/labeled-control.hbs`,
   // Inputs
   COMPONENT_INPUT_TEXTFIELD: `${basePathComponent}/input-textfield/input-textfield.hbs`,
-  COMPONENT_INPUT_DROPDOWN: `${basePathComponent}/input-dropdown/input-dropdown.hbs`,
+  COMPONENT_INPUT_DROPDOWN: `${basePathComponent}/input-choice/input-dropdown/input-dropdown.hbs`,
   COMPONENT_INPUT_NUMBER_SPINNER: `${basePathComponent}/input-number-spinner/input-number-spinner.hbs`,
   COMPONENT_INPUT_TEXTAREA: `${basePathComponent}/input-textarea/input-textarea.hbs`,
   COMPONENT_INPUT_RADIO_BUTTON_GROUP: `${basePathComponent}/input-radio-button-group/input-radio-button-group.hbs`,
